solutions/lab7/requirements.txt
PyYAML>=6.0
jsonschema>=4.0.0 